lab1-distributed-basics/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
requests==2.31.0
